+++ pdqcif/src/test_main.cpp
@@ -8,16 +8,16 @@
 int main(int argc, char** argv)
 {
    auto start = std::chrono::high_resolution_clock::now();
-   //pegtl::file_input in("C:\\data\\ideal_shorter.cif");
-   pegtl::file_input in("C:\\data\\ian.cif");
-   //pegtl::file_input in("C:\\data\\supershort.cif");
-
+   //pegtl::file_input in("C:\\data\\ideal.cif");
+   //pegtl::file_input in("C:\\data\\ian.cif");
+   pegtl::file_input in("C:\\data\\test.cif");
+  
 
    auto file_open = std::chrono::high_resolution_clock::now();
    cif::Cif cif{};
    try {
-       //pegtl::complete_trace<cif::cif>(in);
-      //pegtl::standard_trace< cif::cif >(in);
+      //pegtl::complete_trace<cif::rules::file>(in);
+      //pegtl::standard_trace< cif::rules::cif >(in);
      //pegtl::tracer< pegtl::tracer_traits< true, true, true > > tr(in);
       //tr.parse< rules::cif >(in);
       cif = cif::read_input(in);
